+++ src/main/java/com/github/ilyamurzinov/service/HotelService.java
@@ -15,5 +15,5 @@ public interface HotelService {
 
     public List<Hotel> listHotel();
 
-    public void removeHotel(long id);
+    public void removeHotel(int id);
 }
